Stacks C++
C++ program read in a text file and output the text file in reverse order using stacks. The stacks double in size every time 10 more characters are added.

The output file contains the entire paragraph in reverse order, excluding all punctuations except for an apostrophe. The amount of times the stack double will be displayed on new line.